apps/be/requirements.txt
fastapi==0.115.0
supabase==2.9.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
uvicorn==0.32.0
email-validator==2.1.1
httpx==0.27.2
